Port ElementEditorSlideout to the slideout module
`Craft.ElementEditorSlideout` is now a real TypeScript subclass of
CpScreenSlideout, completing the slideout family; `Craft.ElementEditor`
remains a page-global seam, constructed on load with the slideout's
chrome handed in. All three slideout globals stay `compatify()`-wrapped
so plugin subclasses built on the legacy `.extend()` API keep working.
